修复：修复清空日志问题
